Create RoleView.tsx
RoleView can be passed a Role, and will display the relevant SVG icon and, if `showText` is true, adds the relevant text and a border on the SVG, to be used as a filter button.

IMPORTANT GOTCHA/FOOTGUN: The role icon SVGs are named after their respective IDs, which is how RoleView picks the correct one.