Spring cleaning
Major-ish refactor. Pretty much no functionality has changed, but quite a few lines have been removed. You love to see it.

There is a couple things to be aware of:

The PageContainer, PageNavigator, etc, are now all in main.tsx, meaning they now appear on every page, and don't have to be added manually. This might not be desired in all circumstances, but I think it's an improvement at this point.

Added to TeamActions.ts is getAllTeams, for getting a list of teams for the homepage. It takes a single argument, which is an object that gets converted into a querystring for the request. Currently it just adds the skillSetMask, but because it takes an object, is flexible enough that pagination can be added.