THE BIG ONE
Old functionality: Only able to create new teams, and on creation, the user would be redirected to the homepage, which is not only jarring, but there's no clear relation between the action and the redirect. (Bad UX!)

New:
Register is now a container for RegisterForm, and now what Register does is it loads that user's existing team, then passes it to RegisterForm. Now, when a user has a team already created (or, if they create one while on the page), the create button changes to an update button, and a delete button appears.

I've also added a status bar at the top, to give clear feedback to user submissions. It handles idle state (before a submission), loading state, creation, updating, deletion, and errors, with various colours. A 200ms delay has been added into the useMutate callback just to ensure the status bar transitions background smoothly.

The 240 character back-end limit on description length is now enforced at the front-end, with a live char counter, like Twitter.

Lastly, there are some text/style tweaks. It's hard to get text right, so it's definitely not perfect, but I think it's closer.