Last minute tweaks
getTeam from TeamActions, the only one that uses makeApiRequest and is capable of returning JSON data, now does so.

Tweaked Register to account for this, and to make it so that the delay introduced in useMutate is instead a minimum 200ms, rather than 200ms added onto the request time. So before, if the request was 1sec, then it would become 1.2sec, but now it stays at 1sec. However, if the request time was 0.1sec, it goes up to 0.2sec.