Fully Satisfying ESlint
Doing all the necessary tweaks to satisfy ESlint. There's some weird stuff going on in PageNavigator.tsx I'd like people to have a look at - maybe Dallon would know what's up with it?

I also turned off the no-prototype-builtins rule. It's there to protect against 1: objects with null prototypes and 2: running functions from arbitrary JSON, neither of which we do. And even then, we have good enough error handling that even if we did do the latter, it wouldn't be a problem. The rule only needs to be on for these pretty specific scenarios which we don't/won't use.