Final Implementation, with surprisingly good UX
More or less final implementation of this feature. I removed the "refresh teams" button, since I don't want users to think they have to click it to trigger a load. I also removed "keepPreviousData", so it's clear to the user when a load is occurring (ie. when they interact with the UI). Not sure why that was added in the first place.

I also moved getTeamsList into Home, since that's the only place it's used. It seemed kind of fitting to have it in TeamActions, but ultimately it's unrelated from a code perspective.

Also added overflow-hidden to Team, because Dallon's overly long team description was causing some annoying scroll bars, and I CBA to wait until I implement the redesign to fix it