Refactoring other stuff
Code has been refactored to make use of the new matchif function, as well as being made otherwise more concise.

Also, loading/error handling has been added to the Team Status section (now labelled Team Skills Needed)

matchif works the same as match, but instead of looking for an exact match to a value, it just looks for a truthy value.

There's also matchFunc and matchifFunc now - they work the same as their non-Func counterparts, except you provide a function that returns a return value, instead of providing the return value directly. This is useful when you only want the return expression to be evaluated if it's being returned, which you might want if the return value is dependent on some data being present.