Creating the Status Messenger
The Status Messenger provides APIs for a series of customisable, stylable messages at the bottom of the screen, as a replacement for the status bar on the Register page, but I also figure it'd be useful for other times we want to communicate app state to the user.

This thing completely breaks the way React is supposed to work by calling forceUpdate from outside a component, and having the data that component renders exist outside of it. I'm not sure how else you handle something like this, though.