Reducing issues for colourblind users
Before this commit, selected roles/tabs were indicated more or less only by a change in fill/text colour from white to green. For people with red/green colourblindness, this isn't a helpful indicator. This change puts thicker borders on selected items, as well as adding a background gradient to the selected tab, and refactors the PageNavigator code to be more concise.